+++ src/transactions/types.ts
@@ -147,13 +147,49 @@ export type InputScriptData = {
 };
 
 /**
+ * The data needed to generate a View Function payload
+ */
+export type InputViewFunctionData = {
+  function: MoveFunctionId;
+  typeArguments?: Array<TypeTag | string>;
+  functionArguments?: Array<EntryFunctionArgumentTypes | SimpleEntryFunctionArgumentTypes>;
+  abi?: ViewFunctionABI;
+};
+
+/**
+ * Data needed to generate a view function payload and fetch the remote ABI
+ */
+export type InputViewFunctionDataWithRemoteABI = InputViewFunctionData & { aptosConfig: AptosConfig };
+
+/**
+ * Data needed to generate a view function, with an already fetched ABI
+ */
+export type InputViewFunctionDataWithABI = InputViewFunctionData & { abi: ViewFunctionABI };
+
+/**
+ * Data need for a generic function ABI, both view and entry
+ */
+export type FunctionABI = {
+  typeParameters: Array<MoveFunctionGenericTypeParam>;
+  parameters: Array<TypeTag>;
+};
+
+/**
  * Interface of an Entry function's ABI.
  *
  * This is used to provide type checking and simple input conversion on ABI based transaction submission.
  */
-export type EntryFunctionABI = {
-  typeParameters: Array<MoveFunctionGenericTypeParam>;
-  parameters: Array<TypeTag>;
+export type EntryFunctionABI = FunctionABI & {
+  signers?: number;
+};
+
+/**
+ * Interface of an View function's ABI.
+ *
+ * This is used to provide type checking and simple input conversion on ABI based transaction submission.
+ */
+export type ViewFunctionABI = FunctionABI & {
+  returnTypes: Array<TypeTag>;
 };
 
 /**