+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Handle `Unauthorized` server error
 - Add function to create object address locally
 - Add function to create token object address locally
+- Add signers to entry function ABI for future signature count checking
+- [`Breaking`] Add type-safe view functions with ABI support
 
 # 1.10.0 (2024-03-11)